+++ src/download.ts
@@ -2,10 +2,10 @@
 
 import * as fs from 'fs';
 import * as os from 'os';
-import got from'got';
 import { pipeline } from 'stream';
-import path from 'path';
 import * as vscode from 'vscode';
+import * as got from 'got'; 
+import * as path from 'path';
 
 export const minimum = '3.0.0';